+++ core-tests/e2e-tests/spring/spring-rest-openapi-v3/src/main/kotlin/com/foo/rest/examples/spring/openapi/v3/aiclassification/basic/ACBasicRest.kt
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RestController
 class ACBasicRest {
 
     @GetMapping
-    open fun getData(
+    fun getData(
         @RequestParam("x") x: String?,
         @RequestParam("y") y: Int?,
         @RequestParam("z" )z: Boolean?