+++ core-tests/e2e-tests/spring/spring-rest-openapi-v3/src/main/kotlin/com/foo/rest/examples/spring/openapi/v3/gson/base/GsonRest.kt
@@ -15,7 +15,7 @@ import javax.servlet.http.HttpServletRequest
 class GsonRest {
 
     @PostMapping(consumes = [MediaType.APPLICATION_JSON_VALUE])
-    open fun post(request: HttpServletRequest) : ResponseEntity<String> {
+    fun post(request: HttpServletRequest) : ResponseEntity<String> {
 
         val json = IOUtils.toString(request.inputStream, StandardCharsets.UTF_8)