+++ core/src/main/kotlin/org/evomaster/core/problem/graphql/service/GraphQLBlackBoxModule.kt
@@ -48,11 +48,16 @@ class GraphQLBlackBoxModule(
         bind(object : TypeLiteral<Archive<*>>() {})
                 .to(object : TypeLiteral<Archive<GraphQLIndividual>>() {})
 
+        bind(Archive::class.java)
+            .to(object : TypeLiteral<Archive<GraphQLIndividual>>() {})
+
         bind(object : TypeLiteral<Minimizer<GraphQLIndividual>>(){})
                 .asEagerSingleton()
 
         bind(object : TypeLiteral<Minimizer<*>>(){})
-                .asEagerSingleton()
+            .to(object : TypeLiteral<Minimizer<GraphQLIndividual>>() {})
+            .asEagerSingleton()
+
 
         bind(object : TypeLiteral<FlakinessDetector<GraphQLIndividual>>(){})
             .asEagerSingleton()