+++ core/src/main/kotlin/org/evomaster/core/problem/graphql/service/GraphQLSampler.kt
@@ -23,7 +23,7 @@ class GraphQLSampler : HttpWsSampler<GraphQLIndividual>() {
 
 
     @PostConstruct
-    open fun initialize() {
+    fun initialize() {
 
         log.debug("Initializing {}", GraphQLSampler::class.simpleName)