+++ core/src/test/kotlin/org/evomaster/core/search/service/AdaptiveParameterControlTest.kt
@@ -5,6 +5,7 @@ import com.google.inject.Module
 import com.netflix.governator.guice.LifecycleInjector
 import org.evomaster.core.BaseModule
 import org.evomaster.core.EMConfig
+import org.evomaster.core.search.service.time.SearchTimeController
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeEach