+++ release_notes.md
@@ -2,6 +2,10 @@
 
 Under development in `master` branch.
 
+### New Features
+
+- Each new "phase" (e.g., minimization and security testing) after the search process has now a time-budget which is proportional to the search budget (default 10 percent).
+
 ### Miscellaneous
 
 - Introduced a mascot for EvoMaster: a red-sorcerer black cat with an arrowhead tail. 
@@ -9,6 +13,7 @@ Under development in `master` branch.
 ### Fixed Bugs
 
 - Fixed issue in generated `statistics.csv`, where entries containing `,` commas are now quoted. 
+- Fixed issue in which progress bar was overriding warning messages. 
 
 # Version 5.1.0